+++ backend/runtimes/local/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @cat-factory/local-server
 
+## 0.35.6
+
+### Patch Changes
+
+- Updated dependencies [3135ae8]
+  - @cat-factory/gitlab@0.4.28
+  - @cat-factory/node-server@0.55.3
+  - @cat-factory/server@0.63.3
+
+## 0.35.5
+
+### Patch Changes
+
+- Updated dependencies [39534d6]
+  - @cat-factory/server@0.63.2
+  - @cat-factory/node-server@0.55.2
+
 ## 0.35.4
 
 ### Patch Changes